Refactor server.js into layered architecture (#120)
