+++ src/server/routes/auth.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+const { API_KEY } = require('../middleware/auth');
+
+// Auth status endpoint (always accessible, no auth required)
+router.get('/status', (req, res) => {
+    res.json({ authRequired: !!API_KEY });
+});
+
+module.exports = router;
